requirements.txt
fastmcp
google-genai
httpx
pydantic
python-dotenv
requests
pillow